Fix travis versioning in multi-module run
When setting the version of a multi-module project, the versions:set goal
must be run in the parent module only. Modules inheriting from the parent
will be detected by the plugin and parent references are updated.
maven-versions-plugin 2.2 throws a NPE if child modules inherit the group
id. Therefore downgrade to 2.1, where this does not occur.